reduce duplicate code, got test_wait working
