+++ tests/test_plunger.py
@@ -15,50 +15,13 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import sys
-from contextlib import contextmanager
-from io import StringIO
-
-import py
 import pytest
 import requests_mock
-from pytest import fixture
 from web3 import HTTPProvider, Web3
 
-from plunger.keys import register_key
 from plunger.plunger import Plunger
+from tests.conftest import args, captured_output
 
-
-@contextmanager
-def captured_output():
-    new_out, new_err = StringIO(), StringIO()
-    old_out, old_err = sys.stdout, sys.stderr
-    try:
-        sys.stdout, sys.stderr = new_out, new_err
-        yield sys.stdout, sys.stderr
-    finally:
-        sys.stdout, sys.stderr = old_out, old_err
-
-
-@fixture(scope="session")
-def web3():
-    web3 = Web3(HTTPProvider("http://0.0.0.0:8545"))
-    web3.eth.defaultAccount = "0x6c626f45e3b7aE5A3998478753634790fd0E82EE"
-    register_key(web3, "key_file=tests/data/key1.json,pass_file=/dev/null")
-    register_key(web3, "key_file=tests/data/key2.json,pass_file=/dev/null")
-    assert len(web3.eth.accounts) > 1
-    assert isinstance(web3.eth.accounts[0], str)
-    assert isinstance(web3.eth.accounts[1], str)
-    return web3
-
-
-@fixture
-def datadir(request):
-    return py.path.local(request.module.__file__).join("..").join("data")
-
-
-def args(arguments):
-    return arguments.split()
 
 class TestPlungerUtils:
     @staticmethod
@@ -96,11 +59,6 @@ class TestPlungerUtils:
 
     @staticmethod
     def mock_3_pending_txs_on_jsonrpc(mock, datadir, account: str):
-        # TODO: Hack my json dump to include these tx hashes
-        # mock.get(f"https://unknown.etherscan.io/txsPending?a={account}", text=datadir.join('etherscan').join('3_pending_txs-list.html').read_text('utf-8'))
-        # mock.get(f"https://unknown.etherscan.io/tx/0x124cb0887d0ea364b402fcc1369b7f9bf4d651bc77d2445aefbeab538dd3aab9", text=datadir.join('etherscan').join('3_pending_txs-get1.html').read_text('utf-8'))
-        # mock.get(f"https://unknown.etherscan.io/tx/0x72e7a42d3e1b0773f62cfa9ee2bc54ff904a908ac2a668678f9c4880fd046f7a", text=datadir.join('etherscan').join('3_pending_txs-get2.html').read_text('utf-8'))
-        # mock.get(f"https://unknown.etherscan.io/tx/0x7bc44a24f93df200a3bd172a5a690bec50c215e7a84fa794bacfb61a211d6559", text=datadir.join('etherscan').join('3_pending_txs-get3.html').read_text('utf-8'))
         response = datadir.join('jsonrpc').join('response.json').read_text('utf-8')
         response = response.replace('OUR_ADDRESS', account.upper())
         mock.post(f"http://0.0.0.0:8545/rpc", text=response)
